# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Marketplace Back End Swager API Documentation",
        "description": "Marketplace Back End Swager API Description",
        "contact": {
            "email": "example@google.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0",
        "x-logo": {
            "url": "https://via.placeholder.com/190x90.png?text=MarketPlace"
        }
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Demo API Server"
        }
    ],
    "paths": {
        "/api/v1/product": {
            "get": {
                "tags": [
                    "product"
                ],
                "summary": "Returns Product API response with all product",
                "description": "returns product API response with all product",
                "operationId": "productAll",
                "responses": {
                    "200": {
                        "description": "Successful Operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "405": {
                        "description": "Invalid Input"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            },
            "put": {
                "tags": [
                    "product"
                ],
                "summary": "Update data product",
                "operationId": "productUpdate",
                "requestBody": {
                    "description": "Input data format",
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "product name",
                                        "type": "string"
                                    },
                                    "price": {
                                        "description": "product price",
                                        "type": "string"
                                    },
                                    "quantity": {
                                        "description": "product quantity",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "405": {
                        "description": "Invalid Input"
                    },
                    "422": {
                        "description": "Required field empty"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            },
            "post": {
                "tags": [
                    "product"
                ],
                "summary": "Create new product data",
                "operationId": "productInsert",
                "requestBody": {
                    "description": "Input data format",
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "product name",
                                        "type": "string"
                                    },
                                    "price": {
                                        "description": "product price",
                                        "type": "string"
                                    },
                                    "quantity": {
                                        "description": "product quantity",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "405": {
                        "description": "Invalid Input"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        },
        "/api/v1/product/{pagePaginate}": {
            "get": {
                "tags": [
                    "product"
                ],
                "summary": "Returns Product API response with Pagination",
                "description": "returns product API response with Pagination",
                "operationId": "productPaginate",
                "parameters": [
                    {
                        "name": "pagePaginate",
                        "in": "query",
                        "description": "Page Index",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "405": {
                        "description": "Invalid Input"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        },
        "/api/v1/product/{sortBy}/{sorting}/{filterByColumn}/{searchByColumn}": {
            "get": {
                "tags": [
                    "product"
                ],
                "summary": "Returns Product API response with sort by column, sorting asc or desc, filter by price higher than, and search by column",
                "description": "Returns Product API response with sort by column, sorting asc or desc, filter by price higher than, and search by column",
                "operationId": "productSorting",
                "parameters": [
                    {
                        "name": "sortBy",
                        "in": "query",
                        "description": "Sorting by column",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    },
                    {
                        "name": "sorting",
                        "in": "query",
                        "description": "Sorting by ascending or descending",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    },
                    {
                        "name": "filterByColumn",
                        "in": "query",
                        "description": "Filter by price is higher than or equal",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1000"
                    },
                    {
                        "name": "searchByColumn",
                        "in": "query",
                        "description": "Search By Column",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "product_name, product_price, product_quantity"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "405": {
                        "description": "Invalid Input"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        },
        "/api/v1/product/{id}": {
            "delete": {
                "tags": [
                    "product"
                ],
                "summary": "Delete data product",
                "operationId": "productDelete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "id data",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "0852686e-9d64-4ada-a61b-5fc72fcadb27"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "405": {
                        "description": "Invalid Input"
                    },
                    "422": {
                        "description": "Required field empty"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        },
        "/api/v1/transaction": {
            "post": {
                "tags": [
                    "transaction"
                ],
                "summary": "Create new transaction data",
                "operationId": "productTransaction",
                "requestBody": {
                    "description": "Input data format",
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "product name",
                                        "type": "string"
                                    },
                                    "quantity": {
                                        "description": "product price",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "405": {
                        "description": "Invalid Input"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "api_key": {
                "type": "apiKey",
                "name": "api_key",
                "in": "header"
            },
            "petstore_auth": {
                "type": "oauth2",
                "flows": {
                    "implicit": {
                        "authorizationUrl": "https://petstore.swagger.io/oauth/dialog",
                        "scopes": {
                            "read:pets": "read your API",
                            "write:pets": "modify pets in your account"
                        }
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "product",
            "description": "Everything about your products",
            "externalDocs": {
                "description": "Find out more",
                "url": "https://swagger.io"
            }
        },
        {
            "name": "transaction",
            "description": "Everything about your transactions",
            "externalDocs": {
                "description": "Find out more",
                "url": "https://swagger.io"
            }
        }
    ]
}